refactor idea - auth server
