+++ apps/auth-server/app-init.js
@@ -16,7 +16,7 @@ const currencyFilter              = require('./nunjucks/currency');
 const limitTo                     = require('./nunjucks/limitTo');
 const jsonFilter                  = require('./nunjucks/json');
 const timestampFilter             = require('./nunjucks/timestamp');
-const replaceIdeaVariablesFilter  = require('./nunjucks/replaceIdeaVariables');
+const replaceResourceVariablesFilter  = require('./nunjucks/replaceResourceVariables');
 const flash                       = require('express-flash');
 const expressSession              = require('express-session');
 // const MemoryStore = expressSession.MemoryStore;
