+++ apps/auth-server/public/css/form.less
@@ -29,7 +29,7 @@
 }
 
 
-#idea-form .form-group, #article-form .form-group{
+#resource-form .form-group {
   margin: 70px 0;
 }
