+++ apps/auth-server/public/css/main.css
@@ -7545,7 +7545,7 @@ header.amsterdam-header .header-side-container {
   line-height: 1.2;
   padding-top: 15px;
 }
-#idea-form .form-group,
+#resource-form .form-group,
 #article-form .form-group {
   margin: 70px 0;
 }
